+++ api/lib/database/index.js
@@ -1,6 +1,8 @@
 const { insert, read, query } = require('./helpers')
 const {
+  getAccountNotifications,
   getValidatorsInfo,
+  getValidatorProfile,
   getMaintenance,
   storeNetwork,
   getNetwork,
@@ -20,7 +22,9 @@ const {
   storeNotificationRegistrations,
   getNotificationRegistrations,
   storeAndGetNewSession,
-  getSession
+  getSession,
+  getRewards,
+  getRewardsValidatorHeight
 } = require('./methods')
 
 function database(dbConfig) {
@@ -31,6 +35,7 @@ function database(dbConfig) {
       read: read(dbConfig)(schema),
       query: query(dbConfig),
       getValidatorsInfo: getValidatorsInfo(dbConfig)(schema),
+      getValidatorProfile: getValidatorProfile(dbConfig)(schema),
       getNotifications: getNotifications(dbConfig)(schema),
       getValidatorInfoByAddress: async (validatorId) => {
         const validatorInfo = await getValidatorsInfo(dbConfig)(schema)(
@@ -49,12 +54,15 @@ function database(dbConfig) {
       getMaintenance: getMaintenance(dbConfig)(schema),
       incrementValidatorViews: incrementValidatorViews(dbConfig)(schema),
       getValidatorsViews: getValidatorsViews(dbConfig)(schema),
+      getAccountNotifications: getAccountNotifications(dbConfig)(schema),
       storeUser: storeUser(dbConfig)(schema),
       getUser: getUser(dbConfig)(schema),
       getSession: getSession(dbConfig)(schema),
       storeAndGetNewSession: storeAndGetNewSession(dbConfig)(schema),
       storeStore: storeStore(dbConfig)(schema),
       getStore: getStore(dbConfig)(schema),
+      getRewards: getRewards(dbConfig)(schema),
+      getRewardsValidatorHeight: getRewardsValidatorHeight(dbConfig)(schema),
       storeNotificationRegistrations: storeNotificationRegistrations(dbConfig)(
         schema
       ),