+++ api/lib/fiatvalues-api.js
@@ -91,7 +91,13 @@ class fiatValueAPI {
 
   async pollNewPriceFeed() {
     this.priceFeed = await this.getNewPriceFeed()
-    await this.updateEMoneyExchangeRates()
+    if (
+      EMoneyFiatExchangeRateApi &&
+      EMoneyAPIUrlMainnet &&
+      EMoneyAPIUrlTestnet
+    ) {
+      await this.updateEMoneyExchangeRates()
+    }
     await this.getFiatValuesForAllCoins()
 
     setTimeout(async () => {