+++ api/lib/source/polkadotV0-source.js
@@ -1,6 +1,6 @@
 const BigNumber = require('bignumber.js')
 const BN = require('bn.js')
-const { orderBy, uniqWith } = require('lodash')
+const { orderBy, uniqWith, keyBy } = require('lodash')
 const { stringToU8a, hexToString, hexToU8a } = require('@polkadot/util')
 const { encodeAddress } = require('@polkadot/util-crypto')
 const Sentry = require('@sentry/node')
@@ -124,17 +124,19 @@ class polkadotAPI extends BaseRESTDataSource {
   }
 
   async getTransactionsV2(extrinsics, blockHeight) {
+    const api = await this.getAPI()
     return Array.isArray(extrinsics)
-      ? this.reducers.transactionsReducerV2(
+      ? await this.reducers.transactionsReducerV2(
           this.network,
           extrinsics,
           blockHeight,
-          this.reducers
+          this.db,
+          api
         )
       : []
   }
 
-  async getAllValidators() {
+  async getValidators(blockHeight, fiatCurrency = 'USD') {
     const api = await this.getAPI()
 
     const [allStashAddresses, validatorAddresses] = await Promise.all([
@@ -172,7 +174,7 @@ class polkadotAPI extends BaseRESTDataSource {
     allValidatorsIdentity = JSON.parse(JSON.stringify(allValidatorsIdentity))
 
     // Get annualized validator rewards
-    const expectedReturns = await this.getAllValidatorsExpectedReturns()
+    const expectedReturns = await this.getValidatorsExpectedReturns()
 
     allValidators.forEach((validator) => {
       if (expectedReturns[validator.accountId.toString()]) {
@@ -209,9 +211,23 @@ class polkadotAPI extends BaseRESTDataSource {
         validator.votingPower = 0
       }
     })
-
-    return allValidators.map((validator) =>
-      this.reducers.validatorReducer(this.network, validator)
+    return Promise.all(
+      allValidators.map(async (validator) => {
+        const fiatValuesResponse = await this.fiatValuesAPI.calculateFiatValues(
+          [
+            {
+              amount: validator.tokens,
+              denom: this.network.stakingDenom
+            }
+          ],
+          fiatCurrency
+        )
+        return this.reducers.validatorReducer(
+          this.network,
+          validator,
+          fiatValuesResponse[this.network.stakingDenom]
+        )
+      })
     )
   }
 
@@ -276,7 +292,7 @@ class polkadotAPI extends BaseRESTDataSource {
   //
   // Annualized validator rewards
   //
-  async getAllValidatorsExpectedReturns() {
+  async getValidatorsExpectedReturns() {
     let expectedReturns = []
     let validatorEraPoints = []
     let endEraValidatorList = []
@@ -401,23 +417,7 @@ class polkadotAPI extends BaseRESTDataSource {
         'Rewards are only queryable per height in Polkadot networks'
       )
     }
-    const schema_prefix = this.network.id.replace(/-/, '_')
-    const table = 'rewards'
-    // TODO would be cool to aggregate the rows in the db already, didn't find how to
-    const query = `
-        query {
-          ${schema_prefix}_${table}(where:{address:{_eq: "${delegatorAddress}"}}) {
-            address
-            validator
-            amount
-            denom
-            height
-          }
-        }
-      `
-    const { data } = await this.db.query(query)
-    const dbRewards = data[`${schema_prefix}_${table}`] || [] // TODO: add a backup plan. If it is not in DB, run the actual function
-
+    const dbRewards = (await this.db.getRewards(delegatorAddress)) || []
     const filteredRewards = await this.filterRewards(dbRewards)
 
     const rewards = this.reducers.dbRewardsReducer(
@@ -548,7 +548,7 @@ class polkadotAPI extends BaseRESTDataSource {
       return []
     }
     const stakingProgress = await this.query(`pallets/staking/progress`)
-    const blockHeight = this.getBlockHeight()
+    const blockHeight = await this.getBlockHeight()
     const api = await this.getAPI() // only needed for constants
     const epochDuration = api.consts.babe.epochDuration
     const sessionsPerEra = api.consts.staking.sessionsPerEra
@@ -556,7 +556,7 @@ class polkadotAPI extends BaseRESTDataSource {
     const eraRemainingBlocks = BigNumber(
       stakingProgress.nextActiveEraEstimate
     ).minus(BigNumber(blockHeight))
-    const allUndelegations = stakingLedger.unlocking || []
+    const allUndelegations = stakingLedger.value.unlocking || []
 
     const undelegationsWithEndTime = allUndelegations.map((undelegation) => {
       const remainingEras = undelegation.era - stakingProgress.activeEra
@@ -671,7 +671,8 @@ class polkadotAPI extends BaseRESTDataSource {
       proposer,
       beneficiary: await this.getNetworkAccountInfo(
         proposal.proposal.beneficiary
-      )
+      ),
+      votes: proposal.council[0] ? proposal.council[0].votes : undefined
     }
   }